test for time_limit
